requirements.txt
requests
Pillow
numpy
spidev
RPi.GPIO
